requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
elasticsearch==7.13.1
fastapi==0.70.1
h11==0.12.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
pydantic==1.8.2
python-dateutil==2.8.2
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0

